move walk to filesystem package
